requirements.txt
asgiref==3.6.0
Django==4.2.1
django-environ==0.10.0
django-neomodel==0.1.0
djangorestframework==3.14.0
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
